Deleted the long comment about checks from MD2.
While the comment is useful, it's applicable not just to MD2, but to any
other format whose dissection is based on source code. I don't want to
replicate the comment everywhere, nor to keep it in one dissector but not
others. Ideally, it should become a part of a future set of guidelines for
writing dissectors.