Add dissectors for "ASCIIish" strings
That is, strings in an unknown ASCII-compatible encoding. Often the encoding
of strings in formats is unspecified, in which case all we can do is assume
they're compatible with ASCII and hope for the best. The already-existing
ascii(Z)String dissectors would mark such strings erroneous if they contained
bytes with the high bit set, but they're really not - we just don't know how
to decode them.