Dispose the log view when disposing the main view
Otherwise, the log view is never disposed, and the application just hangs
in the background forever.

I don't move MainView.dispose into View, since the main view should be the
only one explicitly disposed by the presenter. However, it's convenient to
have the implementation of dispose in SwingView, so I do that.