Rename Quality and its members to better reflect their meaning
The main problem with the old names is that Quality.Good doesn't necessarily
mean that the piece has no problems - it merely says that _this_ note doesn't
imply any problems with it. Similarly with other qualities. So it's not really
_quality_, but rather the quality _impact_ of the note.

If I rename Quality to QualityImpact, however, the enumerator names no longer
sound natural. The impact itself can't be "dubious" or "bad", after all.
So I decided to switch to NoteSeverity and use familiar severity levels as
the enumerators instead.

The only problem I see with the new names is that the difference between Error
and Failure is unclear. I try to mitigate that with comments describing the
semantics of each level. Which are overdue, anyway.