Add a soft size limit to DissectionContext
Its purpose is to provide a hint to the dissector where its "territory" ends.
This enables dissectors to behave well when they're used to dissect a
subsection of a file, rather than the whole thing.

For example, if a file format has its header at the end, then the dissector
for that format can use the soft limit to locate said end. Then it will still
work correctly when the file is embedded in a larger file.

There's currently no way to conveniently modify the soft limit - but there
will be.