MD2, MD3: make the file size constraints nonessential
Instead, check the size against the input blob size before applying it.
This way, we can expand past the soft limit, but not the hard one.

Not fond of the repeated complicated minimum calculation... Time will tell
whether this is a common enough pattern to warrant abstraction. Maybe we can
just pass the size from the header directly to builder.fixSize and deal with
possible overflows during postprocessing.